adding another dataset for recommender segmentation exploration
adding the amazon 2012 fine food reviews csv file.
adding a class to read the file and write sorted output of utility matrix columns.
adding a class to read the sorted file into a utility matrix.
editing the test file which explores the use of density based clustering as part of segmentation in recommender systems.